+++ NEWS.md
@@ -5,6 +5,10 @@
 * `cc` uses a new form of correction giving equivariant intervals. Also allow
   consistency with the continuity-corrected McNemar test (or an intermediate 
   correction of the user's choosing).
+* `skew` for skewness correction (under evaluation). 
+* `bcf` for variance bias correction (under evaluation).
+
+
 # ratesci 0.5.0 (2025-01-10)
 
 ## New features